working db submit on create
